+++ src/tt.h
@@ -64,8 +64,8 @@ private:
 
 class TranspositionTable {
 
-  static constexpr size_t CacheLineSize = 64;
-  static constexpr size_t ClusterSize = 3;
+  static constexpr int CacheLineSize = 64;
+  static constexpr int ClusterSize = 3;
 
   struct Cluster {
     TTEntry entry[ClusterSize];
